feat: add `migration:refresh`, `migration:reset`, `migration:fresh` and `db:wipe` commands (#764)
* feat: added migration:reset command

* feat: added shouldCloseConnectionAfterMigrations in MigrationsBase

* feat(commands): added migration:refresh

* test(commands): added migration:refresh tests

* fix: failing test on windows cause of path backslashes

* chore: remove log

* fix: use del-cli instead of del

See : https://github.com/adonisjs/mrm-preset/pull/25

* feat: add `dropAllTables` to QueryClient and TransactionClient

* test: add failling test for exception raised when using `dropAllTables` with no tables in database

* fix: `dropAllTables` should not raises exception when database is empty

* test(commands): add test for db:wipe

* feat(commands): add `db:wipe` command

* feat(commands): add `migration:fresh` command

* test(commands): add test for `migration:fresh`

* chore: update description of new commands `force` flags

* chore: update dryRun flags description of new commands

+++ src/Dialects/Pg.ts
@@ -57,6 +57,8 @@ export class PgDialect implements DialectContract {
    */
   public async dropAllTables(schemas: string[]) {
     const tables = await this.getAllTables(schemas)
+    if (!tables.length) return
+
     await this.client.rawQuery(`DROP table ${tables.join(',')} CASCADE;`)
   }
